+++ package.json
@@ -5,7 +5,7 @@
   "main": "src/index.html",
   "scripts": {
     "test": "grunt travis -v",
-    "workbox": "node updateServiceWorker.js"
+    "workbox": "node updateSW.js"
   },
   "repository": {
     "type": "git",
